Bad log statements no longer clog up the queue
`writePoints` always clears the buffer (`batchP`) and updates
`lastBatchUpdate`, regardless of if there was a write failure.

Since there is no real way to differentiate InfluxDB write errors,
the best that we can do is just attempt a write and then move on.
If it worked, it worked; if it failed, it failed.  No big deal.

Previously, a failed write would cause all subsequent log messages
to stack up indefinitely.  Because the failed message would still
be at the top of the queue, it would continue to fail the entire
batch, and the batch would just get longer and longer and longer.

This obviously caused performance issues (it ate up a lot of memory
over time, and it caused every log statement to attempt to write
to InfluxDB, due to the batch size being reached every time).

This fix cleans all that up.